refactor(docs): split catalog (sync-catalog) from docs-artifact (gen-block-docs); --write is catalog-only
