+++ packages/blocks/package.json
@@ -49,7 +49,7 @@
     "./docs/*": "./docs/*"
   },
   "scripts": {
-    "prebuild": "node ../../scripts/sync-block-docs.mjs --docs-only",
+    "prebuild": "node ../../scripts/gen-block-docs.mjs",
     "build": "tsc --build",
     "test": "node --test dist/conditional-exports.test.js dist/vendorize-map.test.js"
   },
