+++ scripts/sync-catalog.mjs
@@ -3,59 +3,46 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /**
- * Keeps the Building Block catalog and the shipped `docs/` artifact in sync with
- * the per-block READMEs. Three modes:
+ * Keeps the committed Building Block *catalog table* in packages/blocks/README.md
+ * in sync with the per-block READMEs. This script ONLY manages the table between
+ * the `<!-- BEGIN:block-catalog -->` / `<!-- END:block-catalog -->` markers — it
+ * leaves the rest of the README (the static decision-tree prose, etc.) untouched
+ * and does NOT generate the shipped `docs/` artifact. That artifact (per-block
+ * docs/<pkg>/ folders + docs/README.md copy) is produced by
+ * scripts/gen-block-docs.mjs, run as the packages/blocks `prebuild` hook.
+ *
+ * Two modes:
  *
  *   --write  (default; `npm run sync-docs`, run MANUALLY when adding/removing a block)
- *     1. Regenerate ONLY the catalog table between the
- *        `<!-- BEGIN:block-catalog -->` / `<!-- END:block-catalog -->` markers in
- *        packages/blocks/README.md (everything else, incl. the static decision-tree
- *        prose, is preserved).
- *     2. Generate the per-block docs folders under packages/blocks/docs/<pkg>/
- *        (mirror every root *.md except CHANGELOG.md).
- *     3. Write packages/blocks/docs/README.md as a copy of the (now catalog-containing)
- *        packages/blocks/README.md.
+ *     Render the catalog table from the discovered blocks and inject it between
+ *     the markers in packages/blocks/README.md. Nothing else in the README changes.
  *
- *   --check  (CI / PR gate)
- *     Regenerate the catalog table in memory and compare it to what is committed
- *     between the markers in packages/blocks/README.md. Exit 1 with an actionable
- *     message if they differ or the markers are missing; exit 0 if in sync. Writes
- *     nothing.
- *
- *   --docs-only  (build/publish `prebuild` hook)
- *     Generate ONLY the gitignored docs/ artifact — per-block folders + docs/README.md
- *     (a verbatim copy of the committed packages/blocks/README.md). Never modifies the
- *     committed packages/blocks/README.md, so builds don't dirty a tracked file.
+ *   --check  (CI / PR gate; `npm run sync-docs:check`)
+ *     Render the catalog table in memory and compare it to what is committed
+ *     between the markers. Exit 1 with an actionable message if they differ or the
+ *     markers are missing; exit 0 if in sync. Writes nothing.
  *
  * Inclusion rule: every package under packages/ that has a README.md and is not in
- * EXCLUDED. For each included package, mirror every root-level *.md EXCEPT the ones in
- * SKIP_MARKDOWN (CHANGELOG.md) — so block-specific docs are picked up automatically.
+ * EXCLUDED. (The package-discovery logic is intentionally duplicated in
+ * gen-block-docs.mjs — the two scripts are kept independent on purpose.)
  */
 
-import { readFileSync, writeFileSync, mkdirSync, readdirSync, existsSync, rmSync } from 'node:fs';
+import { readFileSync, writeFileSync, readdirSync, existsSync } from 'node:fs';
 import { join, dirname } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const packagesDir = join(__dirname, '..', 'packages');
-const outDir = join(packagesDir, 'blocks', 'docs');
 const readmePath = join(packagesDir, 'blocks', 'README.md');
 
 const EXCLUDED = new Set(['blocks', 'data-common', 'foundations', 'create-blocks-app']);
-
-// Root-level markdown that should NOT be mirrored into the per-block doc folder.
-const SKIP_MARKDOWN = new Set(['CHANGELOG.md']);
 
 const BEGIN_MARKER = '<!-- BEGIN:block-catalog -->';
 const END_MARKER = '<!-- END:block-catalog -->';
 
 const SYNC_HINT = 'Run `npm run sync-docs` and commit the result.';
 
-const mode = process.argv.includes('--check')
-  ? 'check'
-  : process.argv.includes('--docs-only')
-    ? 'docs-only'
-    : 'write';
+const mode = process.argv.includes('--check') ? 'check' : 'write';
 
 const packages = getPackages();
 const catalog = buildCatalog(packages);
@@ -64,7 +51,7 @@ const table = renderCatalogTable(catalog);
 if (mode === 'check') {
   runCheck();
 } else {
-  runGenerate(mode);
+  runWrite();
 }
 
 // ─── Modes ───────────────────────────────────────────────────────────────────
@@ -91,26 +78,11 @@ function runCheck() {
   process.exit(0);
 }
 
-function runGenerate(currentMode) {
+function runWrite() {
   const readme = readFileSync(readmePath, 'utf-8');
-
-  // --write injects the fresh catalog into the committed README; --docs-only must
-  // never touch the committed README, so it copies it verbatim into docs/.
-  let docsReadme = readme;
-  if (currentMode === 'write') {
-    const updated = injectCatalog(readme, table);
-    if (updated !== readme) writeFileSync(readmePath, updated);
-    docsReadme = updated;
-  }
-
-  generatePerBlockDocs();
-  writeFileSync(join(outDir, 'README.md'), docsReadme);
-
-  const where =
-    currentMode === 'write'
-      ? 'packages/blocks/README.md catalog + packages/blocks/docs/'
-      : 'packages/blocks/docs/';
-  console.log(`Synced ${catalog.length} block docs → ${where}`);
+  const updated = injectCatalog(readme, table);
+  if (updated !== readme) writeFileSync(readmePath, updated);
+  console.log(`Synced ${catalog.length} blocks → packages/blocks/README.md catalog`);
 }
 
 // ─── Catalog ─────────────────────────────────────────────────────────────────
@@ -135,29 +107,6 @@ function renderCatalogTable(entries) {
     (e) => `| [${e.pkg}](./docs/${e.pkg}/README.md) | ${e.blurb || '—'} | ${e.keywords || '—'} |`,
   );
   return ['| Block | What it does | Keywords |', '|-------|--------------|----------|', ...rows].join('\n');
-}
-
-// ─── docs/ artifact ──────────────────────────────────────────────────────────
-
-function generatePerBlockDocs() {
-  // Clean and recreate so removed blocks/files don't linger.
-  rmSync(outDir, { recursive: true, force: true });
-  mkdirSync(outDir, { recursive: true });
-
-  for (const pkg of packages) {
-    const pkgDir = join(packagesDir, pkg);
-    const blockOutDir = join(outDir, pkg);
-    mkdirSync(blockOutDir, { recursive: true });
-
-    // Mirror every root-level markdown file (README.md, API.md, DESIGN.md, and any
-    // block-specific docs) except the skipped ones, so new docs ship automatically.
-    const mdFiles = readdirSync(pkgDir, { withFileTypes: true }).filter(
-      (entry) => entry.isFile() && entry.name.endsWith('.md') && !SKIP_MARKDOWN.has(entry.name),
-    );
-    for (const entry of mdFiles) {
-      writeFileSync(join(blockOutDir, entry.name), readFileSync(join(pkgDir, entry.name), 'utf-8'));
-    }
-  }
 }
 
 // ─── Marker helpers ──────────────────────────────────────────────────────────
